Doing 2 task #8

+++ src/main/java/store/services/implementation/ProductCategoryServiceImpl.java
@@ -1,10 +1,13 @@
 package store.services.implementation;
 
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import store.dao.interfaces.ProductCategoryDAO;
+import store.dao.interfaces.ProductDAO;
 import store.dto.ProductCategoryDTO;
+import store.entities.Product;
 import store.entities.ProductCategory;
 import store.exceptions.DAOException;
 import store.services.interfaces.EntityDTOMapper;
@@ -17,7 +20,7 @@ import java.util.stream.Collectors;
  * @author Vadim Popov.
  * PopoWadim@yandex.ru
  **/
-
+@Slf4j
 @Service("productCategoryService")
 public class ProductCategoryServiceImpl implements ProductCategoryService {
 
@@ -25,16 +28,23 @@ public class ProductCategoryServiceImpl implements ProductCategoryService {
     private ProductCategoryDAO productCategoryDAO;
 
     @Autowired
-    EntityDTOMapper entityDTOMapper;
+    private ProductDAO productDAO;
 
+    @Autowired
+    EntityDTOMapper entityDTOMapper;
 
     @Override
     @Transactional
     public void createEntity(ProductCategoryDTO productCategoryDTO) throws DAOException {
-        ProductCategory productCategory = (productCategoryDTO.getProductCategoryId() != 0) ?
-                productCategoryDAO.read(productCategoryDTO.getProductCategoryId()) : new ProductCategory();
+        if (productCategoryDAO.getProductCategoryByName(productCategoryDTO.getName()) != null){
+            DAOException ex = new DAOException("Category with the same name is already exists");
+            log.error("error",ex);
+            throw ex;
+        }
+        ProductCategory productCategory = new ProductCategory();
         entityDTOMapper.mapProductCategoryFromDTO(productCategory, productCategoryDTO);
         productCategoryDAO.create(productCategory);
+        log.info("Creating new Category " + productCategory);
     }
 
     @Override
@@ -46,8 +56,12 @@ public class ProductCategoryServiceImpl implements ProductCategoryService {
     @Override
     @Transactional
     public void updateEntity(ProductCategoryDTO productCategoryDTO) throws DAOException {
-        ProductCategory productCategory = (productCategoryDTO.getProductCategoryId() != 0) ?
-                productCategoryDAO.read(productCategoryDTO.getProductCategoryId()) : new ProductCategory();
+        if (productCategoryDAO.getProductCategoryByName(productCategoryDTO.getName()) != null){
+            DAOException ex = new DAOException("Category with new name is the same like another vendor exists or you are not changing anything");
+            log.error("error", ex.getMessage());
+            throw ex;
+        }
+        ProductCategory productCategory = productCategoryDAO.read(productCategoryDTO.getProductCategoryId());
         entityDTOMapper.mapProductCategoryFromDTO(productCategory, productCategoryDTO);
         productCategoryDAO.update(productCategory);
     }
@@ -58,6 +72,12 @@ public class ProductCategoryServiceImpl implements ProductCategoryService {
         ProductCategory productCategory = (productCategoryDTO.getProductCategoryId() != 0) ?
                 productCategoryDAO.read(productCategoryDTO.getProductCategoryId()) : new ProductCategory();
         entityDTOMapper.mapProductCategoryFromDTO(productCategory, productCategoryDTO);
+        List<Product> productList = productDAO.getAllProductByCategory(productCategoryDTO.getName());
+        if (productList != null && !productList.isEmpty()){
+            DAOException ex = new DAOException("This category contain products, delete or change there before");
+            log.error("error", ex.getMessage());
+            throw ex;
+        }
         productCategoryDAO.delete(productCategory);
     }
